Breaking out work task
